Script descarga/pyproject.toml
[project]
name = "script-descarga"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.4.9",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "sodapy>=2.2.0",
]
